Fix stats multiplier

+++ Assets/Developers/Panta/EntityStatsMultiplierSO.cs
@@ -6,15 +6,15 @@ using UnityEngine;
 public class EntityStatsMultiplierSO : ScriptableObject
 {
     [Header("Defensive Stats")]
-    public int MaxHpMultiplier;
-    public int ArmorMultuplier;
+    public float MaxHpMultiplier;
+    public float ArmorMultuplier;
 
     [Header("Offensive Stats")]
-    public int EngageRangeMultiplier;
-    public int AttackRangeMultiplier;
+    public float EngageRangeMultiplier;
+    public float AttackRangeMultiplier;
     public float AttackSpeedMultiplier;
 
-    public int DamageMultiplier;
+    public float DamageMultiplier;
 
     [Header("Utility")]
     public QuantityOfResources Cost;
